Use a tuple stored in `bot.memory` for known-good patterns
In the spirit of "explicit is better than implicit", prevent accidental
modification of `KNOWN_GOOD` by converting it from `list` to `tuple`.

Is this a Very Practical Change? No, not really; but storing a set of
values that are never expected to change at runtime in a mutable type
rubs me the wrong way.